Refactor handoff output into SRP modules
